Split the LocalDbConfig file into a separate files in the right namespaces (and cleaning them up)
