+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/LocalDb/LocalDbInstanceElement.netfx.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#if NETFRAMEWORK
+
+using System.Configuration;
+
+namespace Microsoft.Data.SqlClient.LocalDb
+{
+    internal sealed class LocalDbInstanceElement : ConfigurationElement
+    {
+        [ConfigurationProperty("name", IsRequired = true)]
+        public string Name => this["name"] as string;
+
+        [ConfigurationProperty("version", IsRequired = true)]
+        public string Version => this["version"] as string;
+    }
+}
+
+#endif
